pyproject.toml
[project]
name = "healther-py"
version = "0.1.0"
description = "Healther check status"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.28.1",
    "sqlmodel>=0.0.22",
    "sqlalchemy>=2.0.32",
    "aiosqlite>=0.20.0",
    "psycopg[binary]>=3.2.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.4.0",
    "email-validator>=2.2.0",
    "redis>=5.0.8",
    "rq>=1.16.2,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
